Moved globally used functions inside wsgate.hpp
