+++ wsgate/wsgate.hpp
@@ -21,12 +21,55 @@
 #define _WSGATE_H_
 
 #include "logging.hpp"
+#include <string>
+using namespace std;
 
 namespace wsgate {
 
     /// Our global logging instance
     typedef wsgate::logger log;
 
+    static const char * const ws_magic = "258EAFA5-E914-47DA-95CA-C5AB0DC85B11";
+
+    //disable two connections to the same host
+    std::map<string, bool> activeConnections;
+
+    int nFormValue(HttpRequest *request, const string & name, int defval) {
+        string tmp(request->FormValues(name).m_sBody);
+        int ret = defval;
+        if (!tmp.empty()) {
+            try {
+                ret = boost::lexical_cast<int>(tmp);
+            } catch (const boost::bad_lexical_cast & e) {  ret = defval; }
+        }
+        return ret;
+    }
+
+    void SplitUserDomain(const string& fullUsername, string& username, string& domain)
+    {
+        std::vector<string> strs;
+        boost::split(strs, fullUsername, boost::is_any_of("\\"));
+        if (strs.size() > 1)
+        {
+            username = strs[1];
+            domain = strs[0];
+        }
+        else
+        {
+            strs.clear();
+            boost::split(strs, fullUsername, boost::is_any_of("@"));
+            if (strs.size() > 1)
+            {
+                username = strs[0];
+                domain = strs[1];
+            }
+            else
+            {
+                username = fullUsername;
+                domain = "";
+            }
+        }
+    }
 }
 
 #endif
