+++ wsgate/wsgate_main.cpp
@@ -17,145 +17,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-#ifdef HAVE_CONFIG_H
-# include "config.h"
-#endif
-
-#ifdef HAVE_WINSOCK2_H
-# include <winsock2.h>
-#endif
-
-#include <ehs/ehs.h>
-#include <vector>
-#include <sstream>
-#include <iostream>
-#include <fstream>
-#include <stdexcept>
-#include <typeinfo>
-#include <cstdlib>
-#include <cstdio>
-#include <cerrno>
-#include <boost/algorithm/string.hpp>
-#include <boost/program_options.hpp>
-#include <boost/filesystem.hpp>
-#include <boost/filesystem/fstream.hpp>
-#include <boost/date_time/posix_time/posix_time.hpp>
-#include <boost/tuple/tuple.hpp>
-#include <boost/regex.hpp>
-#include <boost/optional/optional.hpp>
-#include <boost/property_tree/ptree.hpp>
-#include <boost/property_tree/ini_parser.hpp>
-#include <boost/property_tree/detail/file_parser_error.hpp>
-#include <cpprest/uri.h>
-#include <cpprest/asyncrt_utils.h>
-
-#ifdef HAVE_SIGNAL_H
-# include <signal.h>
-#endif
-#ifdef HAVE_SYS_TYPES_H
-# include <sys/types.h>
-#endif
-#ifdef HAVE_SYS_STAT_H
-# include <sys/stat.h>
-#endif
-#ifdef HAVE_FCNTL_H
-# include <fcntl.h>
-#endif
-#ifdef HAVE_UNISTD_H
- #include <unistd.h>
-#endif
-#ifdef HAVE_SYS_SOCKET_H
-# include <sys/socket.h>
-#endif
-#ifdef HAVE_SYS_WAIT_H
-# include <sys/wait.h>
-#endif
-#ifdef HAVE_SYS_RESOURCE_H
-# include <sys/resource.h>
-#endif
-#ifdef HAVE_NETINET_IN_H
-# include <netinet/in.h>
-#endif
-#ifdef HAVE_ARPA_INET_H
-#include <arpa/inet.h>
-#endif
-
-#include "common.hpp"
-#include "btexception.hpp"
-#include "base64.hpp"
-#include "sha1.hpp"
-#include "logging.hpp"
-#include "wsendpoint.hpp"
+#include "wsgateEHS.hpp"
 #include "wsgate.hpp"
-#include "myrawsocket.hpp"
-#include "nova_token_auth.hpp"
-
-#ifdef _WIN32
-#include <direct.h>
-# define WIN32_LEAN_AND_MEAN
-# include <windows.h>
-# include <shlobj.h>
-# include "NTService.hpp"
-#endif
-
-using namespace std;
-using boost::algorithm::iequals;
-using boost::algorithm::to_lower_copy;
-using boost::algorithm::ends_with;
-using boost::algorithm::replace_all;
-using boost::algorithm::to_upper_copy;
-using boost::algorithm::trim_right_copy_if;
-using boost::algorithm::trim;
-using boost::algorithm::is_any_of;
-using boost::algorithm::split;
-namespace po = boost::program_options;
-namespace fs = boost::filesystem;
-namespace pt = boost::posix_time;
-using boost::filesystem::path;
-using namespace utility::conversions;
 
 namespace wsgate {
-    static const char * const ws_magic = "258EAFA5-E914-47DA-95CA-C5AB0DC85B11";
-
-    //disable two connections to the same host
-    std::map<string, bool> activeConnections;
-
-    int nFormValue(HttpRequest *request, const string & name, int defval) {
-        string tmp(request->FormValues(name).m_sBody);
-        int ret = defval;
-        if (!tmp.empty()) {
-            try {
-                ret = boost::lexical_cast<int>(tmp);
-            } catch (const boost::bad_lexical_cast & e) {  ret = defval; }
-        }
-        return ret;
-    }
-
-    void SplitUserDomain(const string& fullUsername, string& username, string& domain)
-    {
-        std::vector<string> strs;
-        boost::split(strs, fullUsername, boost::is_any_of("\\"));
-        if (strs.size() > 1)
-        {
-            username = strs[1];
-            domain = strs[0];
-        }
-        else
-        {
-            strs.clear();
-            boost::split(strs, fullUsername, boost::is_any_of("@"));
-            if (strs.size() > 1)
-            {
-                username = strs[0];
-                domain = strs[1];
-            }
-            else
-            {
-                username = fullUsername;
-                domain = "";
-            }
-        }
-    }
+    
 
 #ifndef _WIN32
     // Bind helper is not needed on win32, because win32 does not
